Added Foreign Frame Reader
Foreign Frame Reader is an object that loads objects into a mutable frame from a foreign representation – currently a collection of JSON files.
In the future it might support CSV files for a list of objects.

Frame Reader can read multiple JSON-encoded data into a single mutable frame. This allows us to split objects into multiple files not only for convenience but also for (early kind of) composability of designs/models.

- Added Memory.allocateSnapshot to allocate an object snapshot from a foreign record without a structural component
- made structure mutable in ObjectSnapshot
- Added preliminary implementation of FrameReader - reader of foreign representation of a frame