Added Layout command and rewrote persistence
This is slightly overloaded commit that was originally meant to be about "Layout Command", however it surfaced that the implementation of the memory archival was not correct. I started to work on the new memory persistence without committing the early implementation of auto layout. I did not bother to make a git surgery.

- added Layout command - lays out the objects in a circle. will be more sophisticated than that later.
- added MakeshiftMemoryStore and related objects

- added read/write subscript to ForeignRecord
- initialise foreign atom and foreign value from Any
- get Any from foreign atom and foreign value
- renamed allocateID(proposed:) to more accurate allocateID(required:)
- added Memory.fetch(component type)
- added state vector  allValues
- added position initialisation using a point
-
- added more documentation